+++ tests/test/react/plugin.rs
@@ -13,7 +13,7 @@ use bevy::prelude::*;
 
 fn on_entity_insertion(In(entity): In<Entity>, mut rcommands: ReactCommands) -> RevokeToken
 {
-    rcommands.on(entity_insertion::<TestComponent>(entity),
+    rcommands.on_revokable(entity_insertion::<TestComponent>(entity),
             move |world: &mut World| syscall(world, entity, update_test_recorder_with_component)
         )
 }