+++ tests/test/react/resource_reactions.rs
@@ -12,7 +12,7 @@ use bevy::prelude::*;
 
 fn on_resource_mutation(mut rcommands: ReactCommands) -> RevokeToken
 {
-    rcommands.on(resource_mutation::<TestReactRes>(), update_test_recorder_with_resource)
+    rcommands.on_revokable(resource_mutation::<TestReactRes>(), update_test_recorder_with_resource)
 }
 
 fn on_resource_mutation_once(mut rcommands: ReactCommands) -> RevokeToken